refactor(analysis): split git freshness owner
