+++ docs/architecture-consolidation-plan.md
@@ -62,6 +62,7 @@ fixtures:
 | Content complexity | `crates/tokmd-analysis/src/content/complexity.rs` and `crates/tokmd-analysis/src/content/complexity/` | `tests/unit.rs` 1451; production owner modules <=187 | Scoring, nesting, and function-span helpers now live under owner modules; remaining work is mostly test split and aggregation cleanup |
 | Content I/O helpers | `crates/tokmd-analysis/src/content/io.rs` and `crates/tokmd-analysis/src/content/io/` | coordinator 351; tag owner/tests 71 | Keep file reads, text detection, hashing, and entropy helpers in `io.rs` for now; keep tag counting and delimited matching in `io/tags.rs`. Preserve `crate::content::io::*` call sites and the `analysis_content_assets` proof scope. |
 | Analysis orchestration | `crates/tokmd-analysis/src/analysis/mod.rs` and `crates/tokmd-analysis/src/analysis/` | coordinator 152; setup 20; files 53; outputs 22; enricher owners <=160 | Derived/source setup, file collection, optional output storage, and inventory/content/git/semantic/code-quality/effort enrichers now live in owner modules while `analyze` remains the public coordinator. Keep `analysis_orchestration` proof scoped to both the historical `analysis.rs` path and the new `analysis/**` owner paths during the file-to-folder transition |
+| Analysis git report | `crates/tokmd-analysis/src/git/mod.rs`, `git/freshness.rs`, and `git/churn.rs` | coordinator 221; freshness/code-age owner 192; predictive churn owner/tests 174 | Keep export-row matching, hotspot, bus-factor, coupling, and intent assembly in `mod.rs` for now; keep freshness, age buckets, and refresh trend in `freshness.rs`; keep predictive churn in `churn.rs`. Preserve the `analysis_git` proof scope. |
 | Analysis archetype inference | `crates/tokmd-analysis/src/archetype/mod.rs` and `crates/tokmd-analysis/src/archetype/rules.rs` | coordinator 74; rule owner/tests 403 | Keep export-row filtering and path normalization in `mod.rs`; keep ordered Rust/Next.js/container/IaC/Python/Node archetype rules and rule tests in `rules.rs`. Preserve `analysis_security_identity` proof coverage. |
 | Analysis preset grid | `crates/tokmd-analysis/src/grid/mod.rs`, `grid/presets.rs`, and `grid/disabled_feature.rs` | `mod.rs` 10; preset matrix owner 445; disabled-feature warning owner 66 | Keep `mod.rs` as the public barrel; keep preset identity, preset-to-enricher matrix, and plan lookup in `presets.rs`; keep disabled-feature warning text in `disabled_feature.rs`. Preserve the `analysis_orchestration` proof scope. |
 | Analysis API surface | `crates/tokmd-analysis/src/api_surface/mod.rs` and `crates/tokmd-analysis/src/api_surface/` | `mod.rs` 13; report owner 227; symbol dispatcher 56; language scanners <=68; symbol tests 385 | Keep `mod.rs` as a thin coordinator, report aggregation in `report.rs`, source scanning in language owner modules under `symbols`, and leave large integration tests under `api_surface/tests` |
